+++ src/sovereign/constants.py
@@ -1,0 +1,1 @@
+TEMPLATE_CTX_PATH = ".sovereign_context.json"