+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 -----------------
 
 * refactor: cleaning up the import path so that sovereign can be imported without side effects
+* feature: allow accessing template context without generating it from scratch... `from sovereign.runtime import template_context`
 
 0.20.2 20-09-2023
 -----------------